feat(onboarding): refactor step indicator into separate component and integrate into onboarding steps
- Extract StepIndicator component for better modularity and reuse across onboarding steps.
- Update ConfigureStep, DetectStep, InstallStep, and StartGatewayStep to utilize the new StepIndicator, enhancing visual feedback during the onboarding process.
